+++ src/test/java/lotto/model/purchase/PaymentTest.java
@@ -5,7 +5,6 @@ import static lotto.constant.ErrorMessage.INVALID_PAYMENT_UNIT;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import lotto.model.draw.Payment;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
